Attempt to sort BLAST table by genomic location (failed)
